Abstracting common elements of Cartesian, Mercator and Series projections.
