+++ src/test/scala/com/unchartedsoftware/mosaic/core/projection/numeric/CartesianProjectionSpec.scala
@@ -1,13 +1,12 @@
-package com.unchartedsoftware.mosaic.core.projection
+package com.unchartedsoftware.mosaic.core.projection.numeric
 
 import org.scalatest._
 import com.unchartedsoftware.mosaic.core.projection._
-import com.unchartedsoftware.mosaic.util._
 import com.unchartedsoftware.mosaic.core.util.ValueExtractor
 import org.apache.spark.sql.Row
 
-class MercatorProjectionSpec extends FunSpec {
-  describe("MercatorProjection") {
+class CartesianProjectionSpec extends FunSpec {
+  describe("CartesianProjection") {
     describe("#getZoomLevel()") {
       it("should return the first component of a tile coordinate as the zoom level") {
         val extractor = new ValueExtractor[(Double, Double)] {
@@ -15,7 +14,7 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
         val coord = (Math.round(Math.random*100).toInt, 0, 0)
         assert(projection.getZoomLevel(coord) === coord._1)
       }
@@ -28,8 +27,8 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
-        val row = Row(Math.random*170-85, Math.random*360-180)
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
+        val row = Row(Math.random, Math.random)
         intercept[Exception] {
           projection.rowToCoords(row, 2)
         }
@@ -41,9 +40,8 @@ class MercatorProjectionSpec extends FunSpec {
             return None
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
-        val row = Row(Math.random*170-85, Math.random*360-180)
-        assert(projection.rowToCoords(row, 0) === None)
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
+        assert(projection.rowToCoords(Row(Math.random, Math.random), 0) === None)
       }
 
       it("should return None when a row's xCol is outside of the defined bounds") {
@@ -52,7 +50,7 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
         assert(projection.rowToCoords(Row(projection.max._1+1, Math.random), 0) === None)
         assert(projection.rowToCoords(Row(projection.min._1-1, Math.random), 0) === None)
         assert(projection.rowToCoords(Row(projection.max._1, Math.random), 0) === None)
@@ -65,7 +63,7 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
         assert(projection.rowToCoords(Row(Math.random, projection.max._2+1), 0) === None)
         assert(projection.rowToCoords(Row(Math.random, projection.min._2-1), 0) === None)
         assert(projection.rowToCoords(Row(Math.random, projection.max._2), 0) === None)
@@ -78,35 +76,23 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
         //fuzz inputs
         for (i <- 0 until 100) {
-          val row = Row(Math.random*360-180, Math.random*170-85)
+          val row = Row(Math.random, Math.random)
           val coords = projection.rowToCoords(row, 0)
           assert(coords.isDefined)
 
           //check zoom level
           assert(projection.getZoomLevel(coords.get._1) === 0, "check zoom level")
 
-          //compute coordinates and bin for lat/lon manually
-          val lon = row.getDouble(0)
-          val lat = row.getDouble(1)
-          val latRad = (-lat) * (Math.PI / 180);
-          val n = 1 << 0;
-          val howFarX = n * ((lon + 180) / 360);
-          val howFarY = n * (1-(Math.log(Math.tan(latRad) + 1/Math.cos(latRad)) / Math.PI)) / 2
-
-          val x = howFarX.toInt
-          val y = howFarY.toInt
-
-          var xBin = ((howFarX - x)*projection.xBins).toInt
-          var yBin = (projection.yBins-1) - ((howFarY - y)*projection.yBins).toInt
-
           //check coordinates
-          assert(coords.get._1._2 === x, "check coordinates")
-          assert(coords.get._1._3 === y, "check coordinates")
+          assert(coords.get._1._2 === 0, "check coordinates")
+          assert(coords.get._1._3 === 0, "check coordinates")
 
           //check bin
+          val xBin = (row.getDouble(0)*projection.xBins).toInt;
+          val yBin = (projection.yBins-1) - (row.getDouble(1)*projection.yBins).toInt;
           assert(coords.get._2 === (xBin + yBin*projection.xBins), "check bin index for " + row.toString)
         }
       }
@@ -117,35 +103,23 @@ class MercatorProjectionSpec extends FunSpec {
             return Some((r.getDouble(0), r.getDouble(1)))
           }
         }
-        val projection = new MercatorProjection(100, 100, 0, 1, extractor, (-180D, -85D), (180D, 85D))
+        val projection = new CartesianProjection(100, 100, 0, 1, extractor, (0D, 0D), (1D, 1D))
         //fuzz inputs
         for (i <- 0 until 100) {
-          val row = Row(Math.random*360-180, Math.random*170-85)
+          val row = Row(Math.random, Math.random)
           val coords = projection.rowToCoords(row, 1)
           assert(coords.isDefined)
 
           //check zoom level
-          assert(projection.getZoomLevel(coords.get._1) === 1, "check zoom level")
-
-          //compute coordinates and bin for lat/lon manually
-          val lon = row.getDouble(0)
-          val lat = row.getDouble(1)
-          val latRad = (-lat) * (Math.PI / 180);
-          val n = 1 << 1;
-          val howFarX = n * ((lon + 180) / 360);
-          val howFarY = n * (1-(Math.log(Math.tan(latRad) + 1/Math.cos(latRad)) / Math.PI)) / 2
-
-          val x = howFarX.toInt
-          val y = howFarY.toInt
-
-          var xBin = ((howFarX - x)*projection.xBins).toInt
-          var yBin = (projection.yBins-1) - ((howFarY - y)*projection.yBins).toInt
+          assert(projection.getZoomLevel(coords.get._1) === 1)
 
           //check coordinates
-          assert(coords.get._1._2 === x, "check coordinates")
-          assert(coords.get._1._3 === y, "check coordinates")
+          assert(coords.get._1._2 === Math.floor(row.getDouble(0)*2))
+          assert(coords.get._1._3 === Math.floor(row.getDouble(1)*2))
 
           //check bin
+          val xBin = ((row.getDouble(0)*200) % 100).toInt
+          val yBin = (projection.yBins - 1) - ((row.getDouble(1)*200) % 100).toInt
           assert(coords.get._2 === (xBin + yBin*projection.xBins), "check bin index for " + row.toString)
         }
       }
